refactor: move block helpers into module
